Add files via upload
In this project the DNA sequence for human MHC, obtained from NCBI:M_57703.1 is analysed and processed using python. The sequence is first validated, then a combination of functions is applyed to quickly produce reverese complement and transcription results; furthermore, statistical paramerters such as nucleotide counts, GC content/ratio, & ratio of codon use for a given amino acid are calculated. Finally, open reading frames and proteins found within the DNA sequence are elucidated and presented in this consice report.¶